requirements.txt
allure-pytest==2.7.0
allure-python-commons==2.7.0
atomicwrites==1.3.0
attrs==19.1.0
beautifulsoup4==4.8.0
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
idna==2.8
importlib-metadata==0.18
JPype1==0.7.0
more-itertools==7.1.0
numpy==1.16.4
packaging==19.0
pandas==0.25.0
pluggy==0.12.0
py==1.8.0
PyMySQL==0.9.3
pyparsing==2.4.0
pytest==4.5.0
pytest-html==1.20.0
pytest-metadata==1.8.0
pytest-repeat==0.8.0
pytest-rerunfailures==7.0
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
requests-toolbelt==0.9.1
six==1.12.0
soupsieve==1.9.2
urllib3==1.25.3
wcwidth==0.1.7
xlrd==1.2.0
zipp==0.5.2
